tbf: permissions: remove ambiguity on offset==None
Before, passing an offset of `None` to get_permissions() could return a Mask(u64), but the
caller would not know what command numbers that mask represented. This
could lead to subtle bugs if the caller assumed a particular offset
which might work a majority of the time but then fail eventually.

This patch requires an offset so there is no ambiguity.