Remove debug print and make it compile with latest nightly
Removes the `debug_print` from the swervolf board IO, in order that the
`print_full_process` and similar functions can be removed by DCE.

Since the VFE patch on the Rust compiler is based on the latest Rust
master, some feature names changed, so tock has to be updated slightly.

This assumes that a Rust toolchain with the name "stage1" exists that
has the patch for VFE applied.